cmd/preview/deploy_utils.go
// Copyright 2023 The Okteto Authors
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
// http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

package preview

import (
	"errors"
	"fmt"
	"strings"

	"github.com/docker/docker/pkg/namesgenerator"
	"github.com/okteto/okteto/cmd/utils"
	"github.com/okteto/okteto/pkg/env"
	oktetoLog "github.com/okteto/okteto/pkg/log"
	modelUtils "github.com/okteto/okteto/pkg/model/utils"
	"github.com/okteto/okteto/pkg/okteto"
)

var (
	ErrNotValidPreviewScope = errors.New("value is invalid for flag 'scope'. Accepted values are ['global', 'personal']")
)

func optionsSetup(cwd string, opts *DeployOptions, args []string) error {
	if len(args) == 0 {
		opts.name = getRandomName(opts.scope)
	} else {
		opts.name = getExpandedName(args[0])
	}

	var err error
	opts.repository, err = getRepository(cwd, opts.repository)
	if err != nil {
		return err
	}
	opts.branch, err = getBranch(cwd, opts.branch)
	if err != nil {
		return err
	}

	if err := validatePreviewType(opts.scope); err != nil {
		return err
	}

	if opts.deprecatedFilename != "" {
		oktetoLog.Warning("the 'filename' flag is deprecated and will be removed in a future version. Please consider using 'file' flag'")
		if opts.file == "" {
			opts.file = opts.deprecatedFilename
		} else {
			oktetoLog.Warning("flags 'filename' and 'file' can not be used at the same time. 'file' flag will take precedence")
		}
	}
	return nil
}

func validatePreviewType(previewType string) error {
	if !(previewType == "global" || previewType == "personal") {
		return fmt.Errorf("%s %w", previewType, ErrNotValidPreviewScope)
	}
	return nil
}

func getRepository(cwd string, repository string) (string, error) {
	if repository != "" {
		return repository, nil
	}

	oktetoLog.Info("inferring git repository URL")
	return modelUtils.GetRepositoryURL(cwd)
}

func getBranch(cwd, branch string) (string, error) {
	if branch != "" {
		return branch, nil
	}

	oktetoLog.Info("inferring git repository branch")
	return utils.GetBranch(cwd)
}

func getRandomName(scope string) string {
	name := strings.ReplaceAll(namesgenerator.GetRandomName(-1), "_", "-")
	if scope == "personal" {
		username := strings.ToLower(okteto.GetSanitizedUsername())
		name = fmt.Sprintf("%s-%s", name, username)
	}
	return name
}

func getExpandedName(name string) string {
	expandedName, err := env.ExpandEnv(name)
	if err != nil {
		return name
	}
	return expandedName
}

func getPreviewURL(name string) string {
	oktetoURL := okteto.GetContext().Name
	previewURL := fmt.Sprintf("%s/previews/%s", oktetoURL, name)
	return previewURL
}
